碰巧搞了一下设置模块。。。就搞完了综合设置和媒体设置

+++ Presentation/RANSUROTTO.BLOG.Web/Administration/Controllers/CustomerController.cs
@@ -76,7 +76,7 @@ namespace RANSUROTTO.BLOG.Admin.Controllers
 
         public virtual ActionResult List()
         {
-            var defaultRoleIds = new List<long> { _customerService.GetCustomerRoleBySystemName(SystemCustomerRoleNames.Registered).Id };
+            var defaultRoleIds = new List<int> { _customerService.GetCustomerRoleBySystemName(SystemCustomerRoleNames.Registered).Id };
 
             var model = new CustomerListModel
             {
